Move the asciidoc files to a src directory
By moving these files to a separate directory, I can now build
in that directory and install the results in the root of the
project tree.  The theory is that the files in the root can then
be the released versions of the guide.

This required a minor change to the Makefile to add an install
target, and moving a couple of files to src/.

Signed-off-by: Al Stone <ahs3@redhat.com>