Remove invalid restrictive text
In section 5, the document stated that it was the only source of valid
UUIDs for _DSD.  This is not correct.  There is at least one other
defined in the ACPI spec itself, and there could be more in the future.

So, removed the incorrect statements.

It is also recommended that these UUIDs and structures be used, so we
stress that with some rewording, as well.

Signed-off-by: Al Stone <ahs3@redhat.com>